Details panel for Person and Ticket
Rather than a modal that pops up for editing a person/ticket, a new
side panel opens that focuses on the person/ticket's "details." It shows
all of the properties of the ticket/person, but a user can also still
update those fields if they like. This migration to a side panel was done
because I eventually want to remove the edit/delete buttons on the cards
and instead, a user will jsut click on them and that will open the side
panel. The delete functionality will be moved elsewhere.